Move helper files out of routes dir
