+++ backend/src/app.ts
@@ -11,7 +11,7 @@ import { startServer, stopServer } from './lib/server'
 import { ServerSideEvents } from './lib/server-side-events'
 import { ansibleTower } from './routes/ansibletower'
 import { apiPaths } from './routes/apiPaths'
-import { authenticated } from './routes/authenticated'
+import { authenticated } from './lib/authenticated'
 import { configure } from './routes/configure'
 import { events, startWatching, stopWatching } from './routes/events'
 import { liveness } from './routes/liveness'
