+++ backend/src/lib/authenticated.ts
@@ -1,8 +1,8 @@
 /* Copyright Contributors to the Open Cluster Management project */
 import { Http2ServerRequest, Http2ServerResponse } from 'http2'
-import { logger } from '../lib/logger'
-import { respondInternalServerError, unauthorized } from '../lib/respond'
-import { getToken, isAuthenticated } from '../lib/token'
+import { logger } from './logger'
+import { respondInternalServerError, unauthorized } from './respond'
+import { getToken, isAuthenticated } from './token'
 
 export async function authenticated(req: Http2ServerRequest, res: Http2ServerResponse): Promise<void> {
   const token = getToken(req)
