+++ backend/src/routes/search.ts
@@ -7,7 +7,7 @@ import { logger } from '../lib/logger'
 import { notFound } from '../lib/respond'
 import { getAuthenticatedToken } from '../lib/token'
 import { getMultiClusterHub } from '../lib/multi-cluster-hub'
-import { getNamespace, getServiceCACertificate } from './serviceAccountToken'
+import { getNamespace, getServiceCACertificate } from '../lib/serviceAccountToken'
 
 const proxyHeaders = [
   constants.HTTP2_HEADER_ACCEPT,
